refactor(error-handler): rename file
