+++ lib/static/new-ui/components/AssertViewResult/index.tsx
@@ -8,7 +8,7 @@ import {Screenshot} from '@/static/new-ui/components/Screenshot';
 import {ImageLabel} from '@/static/new-ui/components/ImageLabel';
 import {getImageDisplayedSize} from '@/static/new-ui/utils';
 import styles from './index.module.css';
-import {ErrorHandler} from '../../features/error-handling/components/ErrorHandling/ErrorHandling';
+import {ErrorHandler} from '../../features/error-handling/components/ErrorHandling';
 
 interface AssertViewResultProps {
     result: ImageEntity;
